@@ -5,8 +5,8 @@
   "main": "src/server/index.js",
   "scripts": {
     "build": "webpack --config webpack.prod.js",
-    "start": "node ./src/server/index.js",
-    "dev": "concurrently -n client,server 'webpack -w --config webpack.dev.js' 'nodemon -w src/server src/server/index.js'"
+    "start": "node -r esm ./src/server/index.js",
+    "dev": "concurrently -n client,server 'webpack -w --config webpack.dev.js' 'nodemon -w src/server -- -r esm src/server/index.js'"
   },
   "author": "Josh Wilson",
   "license": "ISC",
@@ -14,12 +14,13 @@
     "bcrypt": "^3.0.6",
     "body-parser": "^1.19.0",
     "connect-redis": "^3.4.2",
-    "dotenv": "^6.2.0",
+    "envalid": "^5.0.0",
+    "esm": "^3.2.25",
     "express": "^4.17.1",
     "express-session": "^1.16.2",
     "express-socket.io-session": "^1.3.5",
     "morgan": "^1.9.1",
-    "pg": "^7.11.0",
+    "pg": "^7.12.0",
     "pg-hstore": "^2.3.3",
     "pug": "^2.0.4",
     "react": "^16.8.6",
@@ -42,7 +43,7 @@
     "babel-loader": "^8.0.6",
     "concurrently": "^4.1.1",
     "nodemon": "^1.19.1",
-    "webpack": "^4.37.0",
+    "webpack": "^4.39.1",
     "webpack-cli": "^3.3.6",
     "webpack-merge": "^4.2.1"
   }
