@@ -1,10 +1,8 @@
-const Sequelize = require('sequelize');
-const Promise = Sequelize.Promise;
+import Sequelize from 'sequelize';
+import UserTable from './tables/user';
+import PostTable from './tables/post';
 
-const UserTable = require('./tables/user');
-const PostTable = require('./tables/post');
-
-module.exports = async () => {
+export default async () => {
   const sequelize = new Sequelize(process.env.SQL_URI, {
     logging: false
   });
